refactor(core): split session start modes out
The registry was born with Deep Research as its only member, so it lived in
that module. `bot` is a sibling, not a Deep Research detail, and nobody looks
for Bot permissions in a file named for Deep Research.

Generated-by: Claude Code
