+++ apps/desktop/renderer-architecture.json
@@ -435,10 +435,10 @@
           "./default-runtime-host-operation.js": 1,
           "./locales/settings-test-result-copy.js": 1,
           "./locales/shell-copy.js": 1,
-          "@maka/core/deep-research": 1,
           "@maka/core/llm-connections": 1,
           "@maka/core/permission": 1,
           "@maka/core/session": 1,
+          "@maka/core/session-start-mode": 1,
           "@maka/core/settings": 1,
           "@maka/core/ui-locale": 1,
           "@maka/ui": 1,
@@ -757,7 +757,7 @@
           "./locales/shell-copy.js": 1,
           "./model-connection-errors.js": 1,
           "./session-workspace-errors.js": 1,
-          "@maka/core/deep-research": 1,
+          "@maka/core/session-start-mode": 1,
           "@maka/core/ui-locale": 1,
           "@maka/ui": 1
         },
