+++ packages/core/package.json
@@ -79,6 +79,7 @@
     "./work-board": "./dist/work-board.js",
     "./workhub-creation-intent": "./dist/workhub-creation-intent.js",
     "./deep-research": "./dist/deep-research.js",
+    "./session-start-mode": "./dist/session-start-mode.js",
     "./long-term-memory": "./dist/long-term-memory.js",
     "./local-memory": "./dist/local-memory.js",
     "./web-search": "./dist/web-search.js",
